logic/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
evdev==1.9.2
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
pynput==1.8.1
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wikipedia==1.4.0
